requirements.txt
Django==5.2.8
dj-database-url==2.1.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
gunicorn==21.2.0
djangorestframework==3.15.1
cloudinary==1.39.1
django-cloudinary-storage==0.3.0
python-dotenv==1.0.1
django-extensions==3.2.3
django-filter==23.5
Pillow==10.3.0
cryptography==42.0.5
Werkzeug==3.0.3
pandas==2.2.0 
numpy==1.26.4 
openpyxl==3.1.5
python-dateutil==2.9.0.post0
requests==2.31.0
urllib3==2.0.7
